[#596] Clean up custom domain CTA
* Introduces a dedicated component for the fancy icon
  link, improving code organization and reusability.
* Simplifies the custom domain preview component by
  using the new fancy icon link component.
* Updates styles and translations for a more cohesive
  user experience.
