+++ src/locales/en.json
@@ -616,6 +616,7 @@
   "t-ai-generated-testimonial": "{0}:",
   "t-benefits-of-identity-plus": "{0}: Identity Plus",
   "upgrade-to": "Upgrade to",
+  "upgrade-for-yourdomain": "Upgrade for secrets.yourdomain.com",
   "secret-access-form": "Secret access form",
   "content-status": "Content status",
   "secret-content-copied-to-clipboard": "Secret content copied to clipboard",
